Segregation of responsibilities
