+++ models/classes/LtiProvider/LtiProviderFieldsMapper.php
@@ -22,10 +22,13 @@ declare(strict_types=1);
 
 namespace oat\taoLti\models\classes\LtiProvider;
 
-use oat\tao\model\oauth\DataStore;
+use oat\oatbox\service\ConfigurableService;
 
-class LtiProviderFieldsMapper
+class LtiProviderFieldsMapper extends ConfigurableService
 {
+    public const SERVICE_ID = 'taoLti/LtiProviderFieldsMapper';
+    public const OPTION_MAP = 'map';
+
     public function map(string $rdfUri): ?string
     {
         return $this->getMap() [$rdfUri] ?? null;
@@ -33,20 +36,6 @@ class LtiProviderFieldsMapper
 
     private function getMap(): array
     {
-        //@todo load from configuration
-        return [
-            RdfLtiProviderRepository::LTI_VERSION => ConfigurableLtiProviderRepository::LTI_VERSION,
-            RdfLtiProviderRepository::LTI_TOOL_CLIENT_ID => ConfigurableLtiProviderRepository::LTI_TOOL_CLIENT_ID,
-            RdfLtiProviderRepository::LTI_TOOL_IDENTIFIER => ConfigurableLtiProviderRepository::LTI_TOOL_IDENTIFIER,
-            RdfLtiProviderRepository::LTI_TOOL_NAME => ConfigurableLtiProviderRepository::LTI_TOOL_NAME,
-            RdfLtiProviderRepository::LTI_TOOL_DEPLOYMENT_IDS => ConfigurableLtiProviderRepository::LTI_TOOL_DEPLOYMENT_IDS,
-            RdfLtiProviderRepository::LTI_TOOL_AUDIENCE => ConfigurableLtiProviderRepository::LTI_TOOL_AUDIENCE,
-            RdfLtiProviderRepository::LTI_TOOL_OIDC_LOGIN_INITATION_URL => ConfigurableLtiProviderRepository::LTI_TOOL_OIDC_LOGIN_INITATION_URL,
-            RdfLtiProviderRepository::LTI_TOOL_LAUNCH_URL => ConfigurableLtiProviderRepository::LTI_TOOL_LAUNCH_URL,
-            RdfLtiProviderRepository::LTI_TOOL_JWKS_URL => ConfigurableLtiProviderRepository::LTI_TOOL_JWKS_URL,
-            RdfLtiProviderRepository::LTI_TOOL_PUBLIC_KEY => ConfigurableLtiProviderRepository::LTI_TOOL_PUBLIC_KEY,
-            DataStore::PROPERTY_OAUTH_SECRET => 'secret',
-            DataStore::PROPERTY_OAUTH_KEY => 'key',
-        ];
+        return $this->getOption(self::OPTION_MAP) ?? [];
     }
 }
